requirements.txt
Flask==2.0.2
paho-mqtt==1.6.1
Pillow==8.4.0
qrcode==6.1